Update VS Code launch configs
Excel Desktop: this launch config will launch Excel but the
task pane cannot be debugged using VS Code. Instead, use Edge DevTools
or Visual Studio to attach to the task pane and debug.

Office Online (Edge): this launch config will allow you to debug
the task pane code but requires manual configuration first. Replace
the url with the share link to an Office Online document. After
debugging is started, you will need to upload the manifest to insert
the add-in into the document. You can then set breakpoints in the
source code, and should hit them when the code is executed.