Merge Office addin taskpane (#179)
* update dependencies

* update mocha and file-loader

* add CODEOWNERS file

* convert-to-single-host: delete .github folder

* Update packages to latest versions (#47)

- Also make small change to launch.json to fix issue that was causing 'Debug Tests' config to fail after updating mocha version.

* Update office-addin-scripts packages (#48)

-Verified tests pass locally

* Remove office-toolbox package (#51)

The office-addin-debugging package is being updated to have sideload integrated
rather than require using office-toolbox.

This removes the sideload and unload package.json scripts which are no longer needed.
They could be replaced using a command like:

"npx office-addin-dev-settings sideload manifest.xml -a excel"

however this would require a dev dependency on office-addin-dev-settings. I think it would
be simpler to just use the "start" and "stop" scripts instead. These scripts were required
previously so that "office-addin-debugging start" would know how to sideload, but now that
this functionality is part of the office-addin-dev-settings package, it can import and call
it directly.

* Update all package versions (#54)

- This will ensure we pick up important office-addin-scripts changes as well