require Node v14 / NPM v6 (#113)
Add "engines" to package.json so that Node v14 and NPM v6 is required.

The .npmrc file makes engine-strict=true at the project level, so "npm install" will give an error if Node or NPM is the wrong version.

I also added the registry setting in .npmrc in case the user config specifies a private registry. This ensures that packages only come from the public registry.

The package-lock.json file is updated for NPM v6 (e.g. lockfile version 1).

When convertToSingleHost.js is run, it will remove "engines" from package.json and delete the .npmrc file.