Making HTML not get auto deleted (#266)
Had to override the default behavior of the CleanWebpackPlugin. It normally deletes all files creates by other plugins that are not webpack.
This is the case for out HTML, css and one .xml file. So I made it just clean the files before compiling and not after.

Now you can run npm run build followed by npm start and the project would work.